Some refactoring on root file
