+++ tests/messages/consumed/traction_to_mlwh_aliquot/test_root_traction.py
@@ -19,7 +19,7 @@ from tol_lab_share.constants.input_traction_volume_tracking_message import (
     INSERT_SIZE,
     RECORDED_AT,
 )
-from tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot import TractionToMlwhAliquot
+from tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot import TractionToMlwhAliquot
 from tests.data.example_create_aliquot_in_mlwh_messages import (
     VALID_TRACTION_TO_WAREHOUSE_MESSAGE as VALID_PAYLOAD,
 )
@@ -33,7 +33,7 @@ def subject():
 class TestTractionToMlwhAliquot:
     def test_init_passes_arguments_to_record(self):
         with patch(
-            "tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot.Record.__init__"
+            "tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot.Record.__init__"
         ) as super_init:
             TractionToMlwhAliquot(VALID_PAYLOAD)
 
@@ -41,7 +41,7 @@ class TestTractionToMlwhAliquot:
 
     def test_message_uuid_returns_correct_message_field(self, subject, any):
         with patch(
-            "tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
+            "tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
             "_make_field"
         ) as make_field:
             field = subject.message_uuid
@@ -54,7 +54,7 @@ class TestTractionToMlwhAliquot:
 
     def test_message_create_date_utc_returns_correct_field(self, subject, any):
         with patch(
-            "tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
+            "tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
             "_make_field"
         ) as make_field:
             field = subject.create_date_utc
@@ -67,7 +67,7 @@ class TestTractionToMlwhAliquot:
 
     def test_lims_id_returns_correct_field(self, subject, any):
         with patch(
-            "tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
+            "tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
             "_make_field"
         ) as make_field:
             field = subject.lims_id
@@ -80,7 +80,7 @@ class TestTractionToMlwhAliquot:
 
     def test_lims_uuid_returns_correct_field(self, subject, any):
         with patch(
-            "tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
+            "tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
             "_make_field"
         ) as make_field:
             field = subject.lims_uuid
@@ -93,7 +93,7 @@ class TestTractionToMlwhAliquot:
 
     def test_aliquot_type_returns_correct_field(self, subject, any):
         with patch(
-            "tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
+            "tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
             "_make_field"
         ) as make_field:
             field = subject.aliquot_type
@@ -106,7 +106,7 @@ class TestTractionToMlwhAliquot:
 
     def test_source_type_returns_correct_field(self, subject, any):
         with patch(
-            "tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
+            "tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
             "_make_field"
         ) as make_field:
             field = subject.source_type
@@ -119,7 +119,7 @@ class TestTractionToMlwhAliquot:
 
     def test_source_barcode_returns_correct_field(self, subject, any):
         with patch(
-            "tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
+            "tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
             "_make_field"
         ) as make_field:
             field = subject.source_barcode
@@ -132,7 +132,7 @@ class TestTractionToMlwhAliquot:
 
     def test_sample_name_returns_correct_field(self, subject, any):
         with patch(
-            "tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot.TractionToMlwhAliquot"
+            "tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot.TractionToMlwhAliquot"
             "._make_field"
         ) as make_field:
             field = subject.sample_name
@@ -145,7 +145,7 @@ class TestTractionToMlwhAliquot:
 
     def test_used_by_type_returns_correct_field(self, subject, any):
         with patch(
-            "tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
+            "tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
             "_make_field"
         ) as make_field:
             field = subject.used_by_type
@@ -158,7 +158,7 @@ class TestTractionToMlwhAliquot:
 
     def test_used_by_barcode_returns_correct_field(self, subject, any):
         with patch(
-            "tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
+            "tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
             "_make_field"
         ) as make_field:
             field = subject.used_by_barcode
@@ -171,7 +171,7 @@ class TestTractionToMlwhAliquot:
 
     def test_volume_returns_correct_field(self, subject, any):
         with patch(
-            "tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
+            "tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
             "_make_field"
         ) as make_field:
             field = subject.volume
@@ -184,7 +184,7 @@ class TestTractionToMlwhAliquot:
 
     def test_concentration_returns_correct_field(self, subject, any):
         with patch(
-            "tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
+            "tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
             "_make_field"
         ) as make_field:
             field = subject.concentration
@@ -197,7 +197,7 @@ class TestTractionToMlwhAliquot:
 
     def test_insert_size_returns_correct_field(self, subject, any):
         with patch(
-            "tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
+            "tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
             "_make_field"
         ) as make_field:
             field = subject.insert_size
@@ -210,7 +210,7 @@ class TestTractionToMlwhAliquot:
 
     def test_recorded_at_returns_correct_field(self, subject, any):
         with patch(
-            "tol_lab_share.messages.consumed.traction_to_warehouse.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
+            "tol_lab_share.messages.consumed.traction_to_mlwh_aliquot.traction_to_mlwh_aliquot.TractionToMlwhAliquot."
             "_make_field"
         ) as make_field:
             field = subject.recorded_at
